app/backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Danis <bzakks21@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.104.1"
uvicorn = "^0.24.0.post1"
pydantic-settings = "^2.1.0"
ultralytics = "^8.0.219"
pillow = "^10.1.0"
python-multipart = "^0.0.6"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

